+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/Stacktical/stacktical-developer-toolkit/compare/v1.1.0...v1.2.0) (2021-05-25)
+
+
+### Features
+
+* allowance for contracts ([eb831b3](https://github.com/Stacktical/stacktical-developer-toolkit/commit/eb831b3b31b7449fab458495ca92bf10f6690d34))
+* bootstrap network analytics ([adee94d](https://github.com/Stacktical/stacktical-developer-toolkit/commit/adee94d3a1135e4fef8fdd6503b8098dcb528662))
+* bootstrap period registry ([96e06cd](https://github.com/Stacktical/stacktical-developer-toolkit/commit/96e06cd6f2058933a88f18b40dd0de384a922a7b))
+* bootstrap stake registry ([3f8f980](https://github.com/Stacktical/stacktical-developer-toolkit/commit/3f8f980406821b41acfc7112262d9c6aa258b00c))
+* bootstraping messenger registry ([17c552b](https://github.com/Stacktical/stacktical-developer-toolkit/commit/17c552b03a344d3888cc367bb2245f6e0b6e2ed7))
+* develop not production chainlink ([5c899a1](https://github.com/Stacktical/stacktical-developer-toolkit/commit/5c899a150749e49145f14610919d7153ad8bbc4c))
+* dotend with path ([beea9e5](https://github.com/Stacktical/stacktical-developer-toolkit/commit/beea9e5e6f560f1ed1f4018592bfff8f31abe006))
+* linux network fix ([9fe42cf](https://github.com/Stacktical/stacktical-developer-toolkit/commit/9fe42cf106ce9db350b51c7fb875cb73e10714b5))
+* old value on staking parameters ([89cf4b7](https://github.com/Stacktical/stacktical-developer-toolkit/commit/89cf4b7a94fb766d8953a2d74c7c5b07fcc44d40))
+* service agreement updated in na and semessenger ([a3f397d](https://github.com/Stacktical/stacktical-developer-toolkit/commit/a3f397df8c6116a35c9bedbf013a99ef1e887a65))
+* staking parameters not optional ([7cd0618](https://github.com/Stacktical/stacktical-developer-toolkit/commit/7cd061894c13a4bbc76fbd8d7b4d6ac66a948656))
+* update precoordinator in networkanalytics and semessenger ([2c591bc](https://github.com/Stacktical/stacktical-developer-toolkit/commit/2c591bce244edd9afed9817cffb9f7a2122ebc0d))
+
+
+### Bug Fixes
+
+* alpine for postgres docker compose ([a351852](https://github.com/Stacktical/stacktical-developer-toolkit/commit/a351852b14363de266cefb5fa7dd8b483ac4c62e))
+* custom ganache to 0.0.0.0 ([4c98217](https://github.com/Stacktical/stacktical-developer-toolkit/commit/4c982178fde525da169349311caf514ae47c96c0))
+* preparing nodes before set precoordinator ([d697c71](https://github.com/Stacktical/stacktical-developer-toolkit/commit/d697c71f15b3178db2dd9f8d5222036fb8ffd357))
+
 ## [1.1.0](https://github.com/Stacktical/stacktical-developer-toolkit/compare/v1.0.0...v1.1.0) (2021-05-25)